Add service for payment
