+++ src/components/PayNowForm/PayNoWForm.jsx
@@ -2,8 +2,7 @@
 import {
   CardNumberElement, CardExpiryElement, CardCvcElement, useElements, useStripe,
 } from '@stripe/react-stripe-js';
-import { useDispatch, useSelector } from 'react-redux';
-import { createBooking } from '../../features/bookings/bookingsSlice';
+import createPayment from '../../services/payments';
 import masterCard from '../../assets/master-card-r.png';
 import visa from '../../assets/visa.png';
 import americanExpress from '../../assets/american-express.png';
@@ -14,36 +13,20 @@ import './styles.css';
 const PayNowForm = () => {
   const elements = useElements();
   const stripe = useStripe();
-  const dispatch = useDispatch();
-  const bookings = useSelector((state) => state.bookings);
   const handleSubmit = async (event) => {
     event.preventDefault();
-    const { error, paymentMethod } = await stripe.createPaymentMethod({
+    const { paymentMethod } = await stripe.createPaymentMethod({
       type: 'card',
       card: elements.getElement(CardNumberElement, CardExpiryElement, CardCvcElement),
     });
-    if (error) {
+    const amount = 500 * 100;
+    try {
+      createPayment(paymentMethod, amount);
+    } catch (error) {
       throw new Error(error);
     }
-    const options = {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        paymentMethod,
-        amount: Math.floor(4 * 100),
-      }),
-    };
-    const response = await fetch('http://localhost:8080/api/payments', options);
-    const data = await response.json();
-    console.log(data);
   };
 
-  const makePayment = () => {
-    dispatch(createBooking(bookings));
-    console.log('valor', bookings);
-  };
   return (
     <form className="card-data__form" onSubmit={handleSubmit}>
       <div>
@@ -72,7 +55,7 @@ const PayNowForm = () => {
         </div>
       </div>
       <div className="card-data__btn-cont">
-        <button className="card-data__btn" onClick={makePayment} type="submit">
+        <button className="card-data__btn" type="submit">
           MAKE PAYMENT
         </button>
       </div>
